fix(l10n): ensure the correct language is selected in localizeTimestamp
The test of supportedLanguages.indexOf was mistakenly checking against 0
instead of -1. This meant that the first item in our list of supported
languages, arabic, was never being matched and other, unsupported ones
were being matched.

Furthermore, the accept-languages package assumes it is initialised with
an array of languages that are sorted in order of precedence. This meant
our defaultLanguage option was effectively being ignored because it was
not at the start of the list.

This change fixes both issues and tweaks the tests so that they'll fail
if either regresses in the future.